+++ glue/plugins/ginga_viewer/client.py
@@ -50,6 +50,9 @@ class GingaClient(ImageClient):
     def _update_axis_labels(self):
         pass
 
+    def _update_and_redraw(self):
+        pass
+
     def set_cmap(self, cmap):
         self._canvas.set_cmap(cmap)