+++ glue/qt/tests/test_qtutil.py
@@ -8,7 +8,7 @@ from ...external.qt import QtGui
 from ...external.qt.QtCore import Qt
 from mock import MagicMock, patch
 from ..qtutil import GlueDataDialog
-from ..qtutil import pretty_number, GlueComboBox, PythonListModel
+from ..qtutil import pretty_number, GlueComboBox, PythonListModel, update_combobox
 
 from glue.config import data_factory
 from glue.core import Subset
@@ -452,3 +452,41 @@ class TestListModel(object):
     def test_iter(self):
         m = PythonListModel([1, 2, 3])
         assert list(m) == [1, 2, 3]
+
+
+
+def test_update_combobox():
+    combo = GlueComboBox()
+    update_combobox(combo, [('a', 1), ('b', 2)])
+    update_combobox(combo, [('c', 3)])
+
+
+def test_update_combobox_indexchanged():
+    # Regression test for bug that caused currentIndexChanged to not be
+    # emitted if the new index happened to be the same as the old one but the
+    # label data was different.
+
+    class MyComboBox(GlueComboBox):
+
+        def __init__(self, *args, **kwargs):
+            self.change_count = 0
+            super(MyComboBox, self).__init__(*args, **kwargs)
+            self.currentIndexChanged.connect(self.changed)
+
+        def changed(self):
+            self.change_count += 1
+
+    combo = MyComboBox()
+    update_combobox(combo, [('a', 1), ('b', 2)])
+    update_combobox(combo, [('c', 3)])
+
+    assert combo.change_count == 2
+    assert combo.currentIndex() == 0
+
+    combo = MyComboBox()
+    update_combobox(combo, [('a', 1), ('b', 2)])
+    update_combobox(combo, [('a', 1), ('b', 3)])
+    update_combobox(combo, [('a', 3), ('b', 1)])
+
+    assert combo.change_count == 3
+    assert combo.currentIndex() == 1