+++ glue/qt/tests/test_widget_properties.py
@@ -76,6 +76,10 @@ def test_combo_text():
         tc.co = 'c'
     assert exc.value.args[0] == "Cannot find text 'c' in combo box"
 
+    tc.co = None
+    assert tc.co == None
+    assert tc._combo.currentIndex() == -1
+
 
 def test_text():
 
@@ -235,8 +239,12 @@ def test_connect_current_combo():
     assert combo.currentIndex() == 0
 
     # TODO: should the following not return an error?
-    t.a = 'c'
-    assert combo.currentIndex() == 0
+    with pytest.raises(ValueError) as exc:
+        t.a = 'c'
+    assert exc.value.args[0] == 'c not found in combo box'
+
+    t.a = None
+    assert combo.currentIndex() == -1
 
 
 def test_connect_float_edit():